Add pinch and move gestures handling as event on camera

+++ Assets/Objects/Canvas/TimeToEnd/Scripts/TimeToEndController.cs
@@ -36,7 +36,7 @@ namespace PhotonInMaze.CanvasGame.TimeToEnd {
         }
 
         public override int GetInitOrder() {
-            return InitOrder.Default;
+            return (int)InitOrder.Default;
         }
     }
 }
